+++ packages/web3-eth/src/utils/format_transaction.ts
@@ -24,7 +24,7 @@ import {
 	FormatType,
 } from 'web3-utils';
 import { Transaction } from 'web3-types';
-import { isNullish } from 'web3-validator';
+import { isNullish, ValidationSchemaInput } from 'web3-validator';
 import { TransactionDataAndInputError } from 'web3-errors';
 import { transactionSchema } from '../schemas';
 
@@ -34,6 +34,9 @@ export function formatTransaction<
 >(
 	transaction: TransactionType,
 	returnFormat: ReturnFormat = DEFAULT_RETURN_FORMAT as ReturnFormat,
+	options: { transactionSchema: ValidationSchemaInput | typeof transactionSchema } = {
+		transactionSchema,
+	},
 ): FormatType<TransactionType, ReturnFormat> {
 	let formattedTransaction = mergeDeep({}, transaction as Record<string, unknown>) as Transaction;
 	if (!isNullish(transaction?.common)) {
@@ -42,16 +45,21 @@ export function formatTransaction<
 			formattedTransaction.common.customChain = { ...transaction.common.customChain };
 	}
 
-	formattedTransaction = format(transactionSchema, formattedTransaction, returnFormat);
+	formattedTransaction = format(options.transactionSchema, formattedTransaction, returnFormat);
 
 	if (!isNullish(formattedTransaction.data) && !isNullish(formattedTransaction.input))
 		throw new TransactionDataAndInputError({
 			data: bytesToBuffer(formattedTransaction.data).toString('hex'),
 			input: bytesToBuffer(formattedTransaction.input).toString('hex'),
 		});
-	else if (!isNullish(formattedTransaction.input)) {
-		formattedTransaction.data = formattedTransaction.input;
-		delete formattedTransaction.input;
+	else if (!isNullish(formattedTransaction.data)) {
+		formattedTransaction.input = formattedTransaction.data;
+		delete formattedTransaction.data;
+	}
+
+	if (!isNullish(formattedTransaction.gasLimit)) {
+		formattedTransaction.gas = formattedTransaction.gasLimit;
+		delete formattedTransaction.gasLimit;
 	}
 
 	return formattedTransaction as FormatType<TransactionType, ReturnFormat>;