+++ packages/web3-providers-ws/src/index.ts
@@ -139,23 +139,10 @@ export default class WebSocketProvider<
 	}
 
 	protected _addSocketListeners(): void {
-		this._socketConnection?.addEventListener('message', this._onMessageHandler);
 		this._socketConnection?.addEventListener('open', this._onOpenHandler);
+		this._socketConnection?.addEventListener('message', this._onMessageHandler);
 		this._socketConnection?.addEventListener('close', e => this._onCloseHandler(e));
-		let errorListeners: unknown[] | undefined;
-		try {
-			errorListeners = this._socketConnection?.listeners('error');
-		} catch (error) {
-			// At some cases (at GitHub pipeline) there is an error raised when trying to access the listeners
-			//	However, no need to do take any specific action in this case beside try adding the event listener for `error`
-			this._socketConnection?.addEventListener('error', this._onErrorHandler);
-			return;
-		}
-		// The error event listener may be already there because we do not remove it like the others
-		// 	So we add it only if it was not already added
-		if (!errorListeners || errorListeners.length === 0) {
-			this._socketConnection?.addEventListener('error', this._onErrorHandler);
-		}
+		this._socketConnection?.addEventListener('error', this._onErrorHandler);
 	}
 
 	protected _removeSocketListeners(): void {
@@ -173,10 +160,11 @@ export default class WebSocketProvider<
 			this._reconnect();
 			return;
 		}
-
 		this._clearQueues(event);
 		this._removeSocketListeners();
 		this._onDisconnect(event.code, event.reason);
+		// disconnect was successful and can safely remove error listener
+		this._socketConnection?.removeEventListener('error', this._onErrorHandler);
 	}
 }