+++ packages/web3-providers-ws/CHANGELOG.md
@@ -53,6 +53,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   `main` and `files` entries in `package.json` changed to `lib/` directory from `dist/` (#5739)
 -   Refactor to use common SocketProvider class (#5683)
+-   Legacy Event `close` has been deprecated, superseded by `disconnect` (#5683)
 
 ## [4.0.1-alpha.5]